refactor(ui): extract UIState + selection helpers to ui/state.rs
Moves SortColumn, GroupedRow, ClickAction, ClickableRegions,
UIState (with its 400+ lines of selection/scroll/grouping methods),
ProcessGroupStats, compute_scroll_offset, and compute_grouped_rows
into ui/state.rs (~740 lines). The block was self-contained —
no theme, format, or panel_block references — so the move is
purely mechanical. Public surface preserved via re-exports.
